python/pytest/requirements.txt
Appium-Python-Client==0.24
selenium==3.3.1
pytest
pytest-xdist
requests
